+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - rename insecure option to accessory control
 - added config option to configure a webroot for the ui
 - add a complete https management to settings (#2617) (@donavanbecker)
+- integration with updated homebridge version
 
 ### Other Changes
 
@@ -26,7 +27,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 
 - `@homebridge/hap-client` @ `v3.1.2`
 - `@homebridge/node-pty-prebuilt-multiarch` @ `v0.13.1`
-- `@homebridge/plugin-ui-utils` @ `v2.1.0`
+- `@homebridge/plugin-ui-utils` @ `v2.1.1-beta.0`
 
 ## v5.9.0 (2025-11-01)